RemoteFeedLoader를 모듈로 이동시켰어요.
