+++ iOS-Architecture-Journey/Modules/Features/EssentialFeed/Tests/EssentialFeedTests/RemoteFeedLoaderTests.swift
@@ -6,28 +6,7 @@
 //
 
 import XCTest
-
-class RemoteFeedLoader {
-	
-	let client: HTTPClient
-	let url: URL
-	
-	init(
-		url: URL,
-		client: HTTPClient
-	) {
-		self.url = url
-		self.client = client
-	}
-	
-	func load() {
-		client.get(from: url)
-	}
-}
-
-protocol HTTPClient {
-	func get(from url: URL)
-}
+import EssentialFeed
 
 class HTTPClientSpy: HTTPClient {
